third_party/blink/web_tests/fast/xsl/transformToFragment-XML-declaration-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
Test for bug 10419: XSLTProcessor transformToFragment fails because of an XML declaration.

SUCCESS
